corregidos errores

+++ src/View/Text.cpp
@@ -20,7 +20,6 @@ Text::Text(const std::string& text, const int fontSize, SdlWindow& window) :
 	this->createText(text);
 }
 
-
 Text::Text(const std::string& text, const int fontSize, SdlWindow& window, int r, int g, int b) :
 	textWidth(0),
 	textHeight(0),
@@ -37,7 +36,7 @@ Text::Text(const std::string& text, const int fontSize, SdlWindow& window, int r
 	if (this->font == nullptr) {
 		throw SdlException("Error al cargar la fuente", TTF_GetError());
 	}
-	this->createText(text,r,g,b);
+	this->createText(text, r, g, b);
 }
 
 Text::~Text() {
@@ -57,7 +56,7 @@ Text::~Text() {
 }
 
 void Text::createText(const std::string& text) {
-	SDL_Color fntColor = {RED,GREEN,BLUE};
+	SDL_Color fntColor = {RED, GREEN, BLUE};
 	// SDL_Color fntColor = { 0x0, 0xFF, 0x0 };
 	if (this->surface != nullptr) {
 		SDL_FreeSurface(this->surface);
@@ -79,20 +78,19 @@ void Text::createText(const std::string& text) {
 	                 &(this->textWidth), &(this->textHeight));
 }
 
-
 void Text::createText(const std::string& text, int r, int g, int b) {
-	SDL_Color fntColor = { r,g,b };
+	SDL_Color fntColor = { (Uint8) r, (Uint8) g, (Uint8) b };
 	// SDL_Color fntColor = { 0x0, 0xFF, 0x0 };
 	if (this->surface != nullptr) {
 		SDL_FreeSurface(this->surface);
-	} 
+	}
 	this->surface = TTF_RenderText_Blended(this->font, text.c_str(), fntColor);
 	// SDL_Color bkgColor = { 0x33, 0x33, 0x33 };
 	// this->surface = TTF_RenderText_Shaded(this->font, text.c_str(), fntColor, bkgColor);
 	if (this->surface == nullptr) {
 		throw SdlException("Error al crear superficie", SDL_GetError());
 	}
-	if (this->texture != nullptr) { 
+	if (this->texture != nullptr) {
 		SDL_DestroyTexture(this->texture);
 	}
 	this->texture = SDL_CreateTextureFromSurface(renderer, surface);
@@ -114,7 +112,7 @@ void Text::setText(const std::string& newText) {
 
 void Text::setText(const std::string& newText, int r, int g, int b) {
 	this->inputText = newText;
-	this->createText(this->inputText,r,g,b); 
+	this->createText(this->inputText, r, g, b);
 }
 
 void Text::render(const int x, const int y) {
